feat: creation of a composable and implementation of interfaces + change all project links that use interfaces to the composable
