+++ racksdb/web/ui/src/components/InfrastructureTable.vue
@@ -6,7 +6,7 @@ SPDX-License-Identifier: GPL-3.0-or-later -->
 
 <script setup lang="ts">
 import type { PropType } from 'vue'
-import type { Infrastructure } from '@/views/InfrastructuresView.vue'
+import type { Infrastructure } from '@/composables/RacksDBAPI'
 import { ref } from 'vue'
 
 const rack = ref()
