+++ racksdb/web/ui/src/views/InfrastructuresView.vue
@@ -9,6 +9,7 @@ import { useHttp } from '@/plugins/http'
 import { ref, onMounted } from 'vue'
 import SearchBar from '@/components/SearchBar.vue'
 import type { Ref } from 'vue'
+import type { Infrastructure } from '@/composables/RacksDBAPI'
 
 const infrastructures: Ref<Array<Infrastructure>> = ref([])
 const http = useHttp()
@@ -26,95 +27,6 @@ onMounted(() => {
   getInfrastructure()
 })
 
-export interface Infrastructure {
-  name: string
-  description: string
-  layout: [
-    {
-      rack: string
-      nodes: [NodeEquipment]
-      network: [NetworkEquipment]
-      storage: [StorageEquipment]
-    }
-  ]
-}
-
-export interface NodeEquipment {
-  type: {
-    id: string
-    model: string
-    height: number
-    width: number
-    specs: string
-    cpu: {
-      sockets: number
-      model: string
-      specs: string
-      cores: number
-    }
-    ram: {
-      dimm: number
-      size: number
-    }
-    storage: [
-      {
-        type: string
-        model: string
-        size: number
-      }
-    ]
-    netifs: [
-      {
-        type: string
-        bandwidth: number
-      }
-    ]
-  }
-  rack: string
-  name: string
-  slot: number
-}
-
-export interface NetworkEquipment {
-  type: {
-    id: string
-    model: string
-    height: number
-    width: number
-    netifs: [
-      {
-        type: string
-        bandwidth: number
-        number: number
-      }
-    ]
-  }
-  tags: []
-  rack: string
-  name: string
-  slot: number
-}
-
-export interface StorageEquipment {
-  type: {
-    id: string
-    model: string
-    height: number
-    width: number
-    disks: [
-      {
-        type: string
-        size: number
-        model: string
-        number: number
-      }
-    ]
-  }
-  tags: []
-  rack: string
-  name: string
-  slot: number
-}
 </script>
 
 <template>
